add database uri and console register form data
