+++ src/lib/dbConnect.js
@@ -1,0 +1,22 @@
+import { MongoClient, ServerApiVersion } from 'mongodb';
+export const collectionNameObj = {
+  servicesCollection:"tools",
+  userCollection:"test_user"
+}
+
+export default function dbConnect (collectionName){
+  
+  const uri = process.env.NEXT_PUBLIC_MONGODB_URI
+  // Create a MongoClient with a MongoClientOptions object to set the Stable API version
+  const client = new MongoClient(uri, {
+    serverApi: {
+      version: ServerApiVersion.v1,
+      strict: true,
+      deprecationErrors: true,
+    }
+  });
+  return client.db(process.env.DB_NAME).collection(collectionName)
+}
+
+
+
